[RT #81249]  test fix-ups for VMS
This patch simplifies and corrects the various types of special handling for
VMS in constant.t.  It turns out the easiest and most robust way to handle
the fact that filename case may or may not be preserved is simply to do all
filename comparisons in a case blind fashion.  This is safe to do because
there is no practical possibility of distinct filenames that differ only by
case, especially with the short list of well-know files being considered
here.